Implementación de Crear y Modificar usuario.

+++ JavaFXGeneracionConstancias/src/jfxgeneracionconstancias/modelos/dao/contratos/IUserDAO.java
@@ -9,10 +9,11 @@ import jfxgeneracionconstancias.modelos.pojo.Usuario;
  * @author Daniel García Arcos
  */
 public interface IUserDAO {
-    
-    public int registrarUsuario(Usuario usuario) throws DAOException;
-    public int editarUsuario(Usuario usuario, int numeroPersonal) throws DAOException;
-    public Usuario obtenerUsuario(int numeroPersonal) throws DAOException;
-    public Usuario autenticarUsuario(int numeroPersonal, String contraseña) throws DAOException;
+
+    public long registrarUsuario(Usuario usuario) throws DAOException;
+    public long editarUsuario(Usuario usuario, long numeroPersonal) throws DAOException;
+    public Usuario obtenerUsuario(long numeroPersonal) throws DAOException;
+    public Usuario autenticarUsuario(long numeroPersonal, String contraseña) throws DAOException;
     public ArrayList<Usuario> obtenerUsuarios() throws DAOException;
+
 }
